requirements.txt
ccxtpro
fastapi
pandas
pyarrow
pyyaml
uvicorn
